fixed circular dependency 4
