+++ core/src/main/resources/application-test.properties
@@ -15,3 +15,4 @@ verifyEmailTimeHour=24
 
 kafka.enable=false
 kafka.bootstrap-servers=localhost:9092
+kafka.topic.user.actions=greencity.user.actions
