+++ dao/pom.xml
@@ -23,12 +23,6 @@
 
     <dependencies>
         <dependency>
-            <groupId>greencity</groupId>
-            <artifactId>service-api</artifactId>
-            <version>${project.version}</version>
-            <scope>compile</scope>
-        </dependency>
-        <dependency>
             <groupId>javax.validation</groupId>
             <artifactId>validation-api</artifactId>
             <version>2.0.1.Final</version>
